refactor : Piece 움직임을 인터페이스화 하여 Article 로 변경
